+++ app/src/main/java/com/example/androiddevelopmentapplicationapp/MainActivity.kt
@@ -12,7 +12,7 @@ import androidx.fragment.app.replace
 import com.example.androidapplicationdevelopmentxml.R
 import com.example.androidapplicationdevelopmentxml.databinding.ActivityMainBinding
 import com.example.androiddevelopmentapplicationapp.ui.categories.CategoriesListFragment
-import com.example.androiddevelopmentapplicationapp.ui.recipes.favorite.FavoritesFragment
+import com.example.androiddevelopmentapplicationapp.ui.recipes.favorites.FavoritesFragment
 
 class MainActivity : AppCompatActivity() {
     private var _binding: ActivityMainBinding? = null
